requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
Django==2.2.7
django-crispy-forms==1.10.0
django-vanilla-views==2.0.0
idna==2.10
Pillow==7.2.0
pytz==2020.4
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.26.2
